+++ src/components/App/spec.tsx
@@ -522,7 +522,7 @@ describe('App', () => {
     });
   });
 
-  describe('handleClick', () => {
+  describe('handleCardClick', () => {
     it(`should remove active card state when an active card exists and the App component is
     clicked`, () => {
       const wrapper = mount(<App />);
@@ -541,6 +541,73 @@ describe('App', () => {
 
       wrapper.find(App).simulate('click');
       expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(true);
+    });
+  });
+
+  describe('handleCardMove', () => {
+    it('should print to the console when a hand card is clicked and the mouse is moved', () => {
+      const wrapper = mount(<App />);
+      const mouseMove = new MouseEvent('mousemove');
+      console.log = jest.fn();
+
+      // draw card
+      const deckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // open hand
+      wrapper.find('.hand').at(0).simulate('click');
+      expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+      expect(wrapper.find(Hand).find(Card)).toHaveLength(1);
+
+      // click hand card
+      const cardSelector = wrapper.find('.hand').find('.card').at(0).prop('onMouseDown');
+      if (cardSelector) act(() => cardSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.hand').find('.card').at(0).hasClass('active')).toBe(true);
+
+      window.dispatchEvent(mouseMove);
+      expect(console.log).toBeCalledWith({ 'x': undefined, 'y': undefined });
+    });
+
+    it(`should stop printing to the console when a hand card is clicked, the mouse is moved, and 
+    the mouse is clicked again`, () => {
+      const wrapper = mount(<App />);
+      const mouseMove = new MouseEvent('mousemove');
+      const mouseClick = new MouseEvent('click');
+      console.log = jest.fn();
+
+      // draw card
+      const deckSelector = wrapper.find('.stackedCard').find('div').at(0).prop('onClick');
+      if (deckSelector) act(() => deckSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.drawCardIcon').at(0)).toHaveLength(0);
+
+      // open hand
+      wrapper.find('.hand').at(0).simulate('click');
+      expect(wrapper.find(Hand).find('div').at(0).hasClass('hidden')).toBe(false);
+      expect(wrapper.find(Hand).find(Card)).toHaveLength(1);
+
+      // click hand card
+      const cardSelector = wrapper.find('.hand').find('.card').at(0).prop('onMouseDown');
+      if (cardSelector) act(() => cardSelector(stopPropagationMouseEvent));
+      wrapper.update();
+      expect(wrapper.find('.hand').find('.card').at(0).hasClass('active')).toBe(true);
+
+      // move the mouse
+      window.dispatchEvent(mouseMove);
+      expect(console.log).toBeCalledWith({ 'x': undefined, 'y': undefined });
+      expect(console.log).toHaveBeenCalledTimes(2);
+
+      // move again just to be sure
+      window.dispatchEvent(mouseMove);
+      expect(console.log).toHaveBeenCalledTimes(4);
+
+      // click the mouse, then move again
+      window.dispatchEvent(mouseClick);
+      window.dispatchEvent(mouseMove);
+      expect(console.log).toHaveBeenCalledTimes(4);
     });
   });