+++ docs/changelog.md
@@ -25,6 +25,7 @@ Ends October 14, 2020
 - fix issue where Hand is Full notification would occur after discarding a card and then drawing a card.
 - fix issue where card is highlighted when its x icon is clicked.
 - fix issue where x icon continued to display for active hand cards.
+- add hand card movement event handlers.
 
 ## Sprint 4
 Ends September 16, 2020